Create task.ipynb
This project consisted of three parts: Python in which the database was given. And it was necessary to analyze the results of the experiment and draw a conclusion whether it was worth launching a new payment mechanic for all users.
SQL in which it was necessary to write certain queries. And thirdly, Python, here it was necessary to implement a function that would automatically load data from a file. And the second function, which will history charts by metrics